@@ -1,5 +1,7 @@
 package com.ibm.wala.cast.python.util;
 
+import java.lang.reflect.InvocationTargetException;
+
 public abstract class PythonInterpreter {
 
   static {
@@ -8,19 +10,35 @@ public abstract class PythonInterpreter {
       Class<PythonInterpreter> i4 =
           (Class<PythonInterpreter>)
               Class.forName("com.ibm.wala.cast.python.jep.CPythonInterpreter");
-      setInterpreter(i4.newInstance());
+      setInterpreter(i4.getDeclaredConstructor().newInstance());
     } catch (ClassNotFoundException
         | InstantiationException
         | IllegalAccessException
-        | UnsatisfiedLinkError e2) {
+        | UnsatisfiedLinkError
+        | IllegalArgumentException
+        | InvocationTargetException
+        | NoSuchMethodException
+        | SecurityException e2) {
       try {
         Class<?> i3 = Class.forName("com.ibm.wala.cast.python.util.Python3Interpreter");
-        setInterpreter((PythonInterpreter) i3.newInstance());
-      } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e) {
+        setInterpreter((PythonInterpreter) i3.getDeclaredConstructor().newInstance());
+      } catch (ClassNotFoundException
+          | InstantiationException
+          | IllegalAccessException
+          | IllegalArgumentException
+          | InvocationTargetException
+          | NoSuchMethodException
+          | SecurityException e) {
         try {
           Class<?> i2 = Class.forName("com.ibm.wala.cast.python.util.Python2Interpreter");
-          setInterpreter((PythonInterpreter) i2.newInstance());
-        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e1) {
+          setInterpreter((PythonInterpreter) i2.getDeclaredConstructor().newInstance());
+        } catch (ClassNotFoundException
+            | InstantiationException
+            | IllegalAccessException
+            | IllegalArgumentException
+            | InvocationTargetException
+            | NoSuchMethodException
+            | SecurityException e1) {
           assert false : e.getMessage() + ", then " + e1.getMessage();
         }
       }
